refactor category views into different files
