+++ api/__init__.py
@@ -73,8 +73,10 @@ def index():
 
 
 from api.auth.views import auth_blueprint
-from api.categories.views import category_blueprint
+from api.categories.categories_views import category_blueprint
+from api.categories.single_category_views import single_category_blueprint
 from api.recipes.views import recipe_blueprint
 app.register_blueprint(auth_blueprint)
 app.register_blueprint(category_blueprint)
+app.register_blueprint(single_category_blueprint)
 app.register_blueprint(recipe_blueprint)
